# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.Code:
    enum:
    - 0
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 2001
    - 2002
    - 2003
    - 2004
    - 3001
    - 3002
    - 3003
    - 3004
    - 4001
    - 4002
    - 4003
    - 4004
    type: integer
    x-enum-varnames:
    - Success
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrInvalidParams
    - ErrBindJSON
    - ErrBindQuery
    - ErrBindForm
    - ErrDatabase
    - ErrRecordNotFound
    - ErrRecordExists
    - ErrRecordInvalidate
    - ErrTokenInvalid
    - ErrTokenExpired
    - ErrTokenMissing
    - ErrPermissionDenied
...
  model.Task:
    properties:
      created_at:
        type: string
      error:
        description: 错误信息
        type: string
      id:
        type: integer
      max_retry:
        description: 最大重试次数
        type: integer
      name:
        description: 任务名称
        type: string
      payload:
        description: 任务载荷（JSON）
        type: string
      retry_count:
        description: 重试次数
        type: integer
      status:
        description: 任务状态
        type: string
      task_id:
        description: 任务ID（UUID）
        type: string
      type:
        description: '任务类型: async, scheduled'
        type: string
      updated_at:
        type: string
      user_id:
        description: 关联用户ID（可选）
        type: integer
    type: object
  response.Response:
    properties:
      code:
        $ref: '#/definitions/errors.Code'
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/cccvno1/nova
  description: |-
    Nova 是一个基于 Go + Echo 的现代化 Web 应用框架，提供用户认证、RBAC 权限管理、文件上传、异步任务队列、审计日志等功能。

    主要特性：
    - 用户认证与授权（JWT）
    - RBAC 权限管理（基于 Casbin）
    - 文件上传与管理（支持本地/OSS/S3）
    - 异步任务队列（基于 Redis）
    - 定时任务调度（基于 Cron）
    - 审计日志（自动记录所有操作）
    - 限流保护（IP/用户维度）
    - 健康检查与监控
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Nova API
  version: "1.0"
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: 查询任务列表，支持按状态和类型过滤，默认查询待处理任务
      parameters:
      - description: 任务状态
        enum:
        - pending
        - processing
        - success
        - failed
        in: query
        name: status
        type: string
      - description: 任务类型
        in: query
        name: type
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 任务列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Task'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取任务列表
      tags:
      - 任务管理
  /tasks/{id}:
    get:
      consumes:
      - application/json
      description: 根据数据库ID获取任务的详细信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 任务详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Task'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取任务详情
      tags:
      - 任务管理
  /tasks/stats:
    get:
      consumes:
      - application/json
      description: 获取各状态任务的数量统计
      produces:
      - application/json
      responses:
        "200":
          description: 统计信息(包含pending, processing, success, failed, total)
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取任务统计
      tags:
      - 任务管理
  /tasks/task/{taskId}:
    get:
      consumes:
      - application/json
      description: 根据业务任务ID（task_id字段）查询任务信息
      parameters:
      - description: 业务任务ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Task'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 根据任务ID查询
      tags:
      - 任务管理
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"